# Conflicts:
#	.idea/dataSources.ids
#	.idea/dataSources.xml
#	.idea/palencia.iml
#	app/Entities/Comunidades.php
#	app/Entities/Cursillos.php
#	app/Entities/SolicitudesEnviadasCursillos.php
#	app/Http/Controllers/PdfController.php
#	public/js/publico/modal.js
#	resources/views/pdf/listarComunidades.blade.php

+++ resources/views/comunidades/index.blade.php
@@ -4,19 +4,19 @@
 @endsection
 @section('contenido')
     <div class="spinner"></div>
-    <div class="hidden table-size-optima altoMaximo">
+    <div class="hidden table-size-optima">
         @if (Auth::check())
-            <div class="row ">
-                @include('comunidades.parciales.buscar')
-            </div>
+            @include('comun.plantillaBuscarIndex',['htmlTemplate'=>'comunidades.parciales.buscar'])
+            @include('comun.plantillaOperacionesIndex',['tabla'=>'comunidades','accion'=>'Nueva'])
             @if(!$comunidades->isEmpty())
                 @foreach ($comunidades as $comunidad)
                     <div>
                         <table class="table-viaoptima table-striped">
-                            <caption class="@if(!$comunidad->activo) foreground-disabled @endif">
-                                {!! $comunidad->comunidad !!}
-                            </caption>
                             <thead>
+                            <tr class="row-fixed">
+                                <th class="tabla-ancho-columna-texto"></th>
+                                <th></th>
+                            </tr>
                             <tr @if(!$comunidad->activo) class="background-disabled" @endif>
                                 <th colspan="2" class="text-right">
                                     <a title="Mostrar"
@@ -32,30 +32,53 @@
                                         </i>
                                     </a>
                                     @if ((Auth::user()->roles->peso)>=config('opciones.roles.administrador')){{--Administrador --}}
-                                    {!! FORM::open(array('route' => array('comunidades.destroy',
-                                    $comunidad->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar')))!!}
-                                    <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
-                                            @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                            class="pull-right lanzarModal"
-                                            data-title="BORRADO"
-                                            data-descripcion="¿Seguro que deseas eliminar esta comunidad?<br><h3><strong>{{$comunidad->comunidad}}</strong></h3>"
-                                            data-footer="true"
-                                            @endif >
-                                        <i class='glyphicon glyphicon-trash full-Width'>
-                                            <div>Borrar</div>
-                                        </i>
-                                    </button>
-                                    @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                        @include ("comun.plantillaBorrado")
+                                    @if($comunidad->activo)
+                                        {!! FORM::open(array('route' => array('comunidades.destroy',
+                                        $comunidad->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar')))!!}
+                                        <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
+                                                @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                                class="pull-right lanzarModal simpleModal"
+                                                data-modal_centro_pantalla="true"
+                                                data-modal_en_la_derecha="false"
+                                                data-modal_sin_etiqueta="true"
+                                                data-modal_ancho="330"
+                                                data-modal_cabecera_color_fondo='rgba(255,0,0,.9)'
+                                                data-modal_cabecera_color_texto='#ffffff'
+                                                data-modal_cuerpo_color_fondo='rgba(255,255,255,1)'
+                                                data-modal_cuerpo_color_texto='"#ffffff'
+                                                data-modal_pie_color_fondo='#400090'
+                                                data-modal_pie_color_texto='"#ffffff'
+                                                data-modal_posicion_vertical="220"
+                                                data-titulo="BORRAR"
+                                                data-pie="true"
+                                                data-descripcion="¿Seguro que deseas eliminar esta comunidad?<br><h3><strong>{{$comunidad->comunidad}}</strong></h3>"
+                                                data-footer="true"
+                                                @endif >
+                                            <i class='glyphicon glyphicon-trash full-Width'>
+                                                <div>Borrar</div>
+                                            </i>
+                                        </button>
+                                        @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                            @include ("comun.plantillaBorrado")
+                                        @endif
+                                        {!! FORM::close() !!}
                                     @endif
-                                    {!! FORM::close() !!}
                                     @endif
+                                </th>
+                            </tr>
+                            <tr>
+                                <th colspan="2" class="cabecera">
+                                    <div class="ellipsis text-center @if(!$comunidad->activo) foreground-disabled @endif"
+                                         @if($comunidad->activo==1) style="background-color:
+                                         {{$comunidad->colorFondo}} !important; color:{{$comunidad->colorTexto}} !important; @endif ">
+                                        {!! $comunidad->comunidad !!}
+                                    </div>
                                 </th>
                             </tr>
                             </thead>
                             <tbody @if(!$comunidad->activo) class="foreground-disabled" @endif>
                             <tr>
-                                <td class="table-autenticado-columna-1">Secretariado:</td>
+                                <td>Secretariado:</td>
                                 <td>
                                     {!! $comunidad->tipo_secretariado !!}
                                 </td>
@@ -65,7 +88,7 @@
                                 <td> @if ($comunidad->esPropia) Si @else No @endif </td>
                             </tr>
                             <tr>
-                                <td>País:</td>
+                                <td>Pa&iacute;s:</td>
                                 <td>
                                     {!! $comunidad->pais !!}
                                 </td>
@@ -89,24 +112,18 @@
                                 </td>
                             </tr>
                             <tr>
-                                <td>Dirección:</td>
+                                <td>Direcci&oacute;n:</td>
                                 <td>
                                     {!! $comunidad->direccion !!}
                                 </td>
                             </tr>
                             <tr>
-                                <td>Comunicación preferida:</td>
+                                <td>Comunicaci&oacute;n preferida:</td>
                                 <td>{{$comunidad->comunicacion_preferida}}</td>
                             </tr>
                             <tr>
                                 <td>Colabora:</td>
                                 <td> @if ($comunidad->esColaborador) Si @else No @endif </td>
-                            </tr>
-                            <tr>
-                                <td>Color Cursos:</td>
-                                <td>
-                                    <div class="ponerCirculoColor" style="background-color:{{$comunidad->color}}"></div>
-                                </td>
                             </tr>
                             <tr>
                                 <td>Activo:</td>
@@ -116,6 +133,7 @@
                         </table>
                     </div>
                 @endforeach
+                {!! $comunidades->appends(Request::only(['comunidad','esPropia','pais','esActivo']))->render()!!}
             @else
                 <div class="">
                     <div class="alert alert-info" role="alert">
@@ -123,12 +141,6 @@
                     </div>
                 </div>
             @endif
-            <div class="row paginationBlock">
-                {!! $comunidades->appends(Request::only(['comunidad','pais']))->render()
-                !!}{{-- Poner el paginador --}}
-                {{--{!! $comunidades->appends(Request::only(['comunidad','esPropia','secretariado','pais']))->render()
-                !!}--}}{{-- Poner el paginador --}}
-            </div>
         @else
             @include('invitado')
         @endif
